Vague 4b — les 14 questions du moteur
Les 14 questions du §4 de la spécification métier, chacune adossée à sa formule,
jamais à une interprétation libre. Une seule route, GET /api/questions, avec les
mêmes paramètres de période que le tableau de bord. Aucune migration : la
Vague 4a avait posé les données manquantes.

La règle qui gouverne tout le moteur : une question sans données répond
« indisponible » avec sa raison, jamais zéro. Une entreprise qui n'a renseigné
aucun coût ne lit pas « produit le plus rentable : T-shirt, 0 % » mais « aucun
produit vendu n'a de coût de revient renseigné », et sait quoi faire pour
obtenir la réponse. Un indicateur faux coûte plus cher qu'un indicateur absent.

Chaque réponse porte son renvoi de formule (§3.6, §3.9…) : on remonte d'un
chiffre affiché à la règle qui l'a produit sans lire le code.

Écart n° 3 clos. L'arbitrage confirme que le code avait raison pour les périodes
ancrées au calendrier : le 8 août, « ce mois » se compare au 1–8 juillet, pas
aux huit derniers jours de juillet. Un point restait divergent et il est
corrigé : une période ancrée en cours s'arrête désormais à aujourd'hui — « ce
mois » vaut le mois à date. Sans quoi la série journalière traînait des jours
futurs à zéro et l'en-tête annonçait « du 1er au 31 août » un 8 août. Une
période personnalisée n'est jamais tronquée : l'utilisateur a choisi ses bornes.

Deux défauts trouvés pendant les tests.

sum() sur un bigint rend un numeric en Postgres, que node-postgres livre sous
forme de CHAÎNE. Les comparaisons coercent en silence, donc le défaut restait
invisible jusqu'à la première soustraction, qui levait « Cannot mix BigInt and
other types ». Chaque sum() est recastée en ::bigint — dans le moteur de
questions et dans celui du tableau de bord, où le même piège dormait sans avoir
encore mordu.

Le vocabulaire sectoriel cassait les accords : « Quel {nom} est le plus
rentable ? » donnait « Quel prestation est le plus rentable ? ». Les libellés
sont désormais écrits en entier par secteur, ce qui supprime la classe entière
de fautes. Seuls les mots changent : aucune formule, aucun seuil, aucun
classement n'est modifié.

Vérifié : typecheck, 307 tests, et 36 contrôles de bout en bout sur la vraie
base — le cas de référence §7 saisi intégralement par l'API puis relu. CA
315,00 € (+12,5 %), bénéfice −60,00 € avec écart de −80,00 € et aucun
pourcentage, panier 31,50 € (+1,3 %), Casquette la plus vendue, Pull le plus de
CA, Casquette la plus rentable à 66,7 % avec le Pull exclu, marge globale
138,00 €, Pull et Sac ex æquo aux moins vendus, Awa Diop en tête sans qu'aucun
« anonyme » n'apparaisse. Tables revérifiées à zéro ligne.

+++ server/src/modules/kpi/depot.ts
@@ -9,6 +9,13 @@ import type { DepenseAgregable, ProduitAgrege, VenteAgregable } from "../../doma
  * et l'index partiel `ventes_kpi_idx` correspond exactement à ce prédicat. Le
  * **calcul** est fait dans `domaine/kpi.ts`, où il se teste au centime.
  *
+ *
+ * **Chaque `sum()` est explicitement recastée en `::bigint`.** Postgres rend
+ * un `numeric` quand on somme des `bigint`, et node-postgres livre un
+ * `numeric` sous forme de CHAINE. La premiere soustraction leve alors
+ * « Cannot mix BigInt and other types » — alors que les comparaisons, elles,
+ * coercent en silence : le defaut se cache jusqu'au premier calcul.
+ *
  * Seule exception : `topProduits` est agrégé en SQL. Remonter toutes les lignes
  * de vente d'une année pour les sommer en mémoire serait absurde, et
  * `GROUP BY` est précisément le travail d'une base.
@@ -96,7 +103,7 @@ export function creerDepotKpi(pool: Pool): DepotKpi {
           pool.query<{ libelle: string; quantite: string; montant: bigint }>(
             `SELECT l.libelle,
                     sum(l.quantite)::text  AS quantite,
-                    sum(l.montant_mineur)  AS montant
+                    sum(l.montant_mineur)::bigint AS montant
                FROM lignes_vente l
                JOIN ventes v ON v.id = l.vente_id
               WHERE v.entreprise_id = $1
